Applying a patch from Tais M. Hansen <tais.hansen@osd.dk> to add initial sys-devel/icecream cluster compiler support for bug #200095.  This is catalyst 2.0.5_pre6 for testing.
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/catalyst/trunk@1263 d1e1f19c-881f-0410-ab34-b69fee027534

+++ targets/stage3/stage3-preclean-chroot.sh
@@ -15,4 +15,10 @@ then
 	cleanup_distcc
 fi
 
+if [ -n "${clst_ICECREAM}" ]
+then
+	run_emerge -C sys-devel/icecream || exit 1
+	cleanup_icecream
+fi
+
 rm -f /var/log/emerge.log
